+++ bigmler/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '1.15.3'
+__version__ = '1.15.4'
 
 if __name__ == '__main__':
     import sys