+++ bigmler/defaults.py
@@ -144,7 +144,9 @@ FLAGS = {
         {'flag': 'dry_run', 'type': 'boolean'},
         {'flag': 'anomaly_tag', 'type': 'string'},
         {'flag': 'anomaly_score_tag', 'type': 'string'},
-        {'flag': 'fast', 'type': 'boolean'}],
+        {'flag': 'fast', 'type': 'boolean'},
+        {'flag': 'project', 'type': 'string'},
+        {'flag': 'project_id', 'type': 'string'}],
     'BigMLer analyze': [
         {'flag': 'k-fold', 'type': 'integer'},
         {'flag': 'cv', 'type': 'boolean'},